+++ apps/admin/src/app/[locale]/(dashboard)/seasons/components/create-season-dialog.tsx
@@ -19,7 +19,7 @@ import CloudUploadIcon from '@mui/icons-material/CloudUpload';
 import { FileUpload, FormikTextField } from '@lems/shared';
 import { AdminSeasonResponseSchema } from '@lems/types/api/admin';
 import { apiFetch } from '../../../../../lib/fetch';
-import { DialogComponentProps } from '../../dialog-provider';
+import { DialogComponentProps } from '../../components/dialog-provider';
 
 interface SeasonFormValues {
   slug: string;
