+++ apps/admin/src/app/[locale]/(dashboard)/teams/components/create-team-button.tsx
@@ -3,7 +3,7 @@
 import { useTranslations } from 'next-intl';
 import { IconButton, Tooltip } from '@mui/material';
 import AddRoundedIcon from '@mui/icons-material/AddRounded';
-import { useDialog } from '../../dialog-provider';
+import { useDialog } from '../../components/dialog-provider';
 import { CreateTeamDialog } from './create-team-dialog';
 
 export const CreateTeamButton = () => {
